// src/app/shared/models/customer-filter-result.ts
export interface CustomerFilterResult {
  companyId: number;
  customerCode: string;
  customerName: string;
  customerId: number;
  zoneId: number;
  zoneName: string;
  salesExecutiveGroupId: number;
  salesExecutiveName: string;
  salesExecutiveId: number;
  customerStatusName: string;
  customerStatusId: number;
}